+++ src/main/java/com/yapp/ios2/fitfty/domain/user/UserService.java
@@ -1,0 +1,5 @@
+package com.yapp.ios2.fitfty.domain.user;
+
+public interface UserService {
+
+}
