+++ src/main/java/com/yapp/ios2/fitfty/domain/user/UserServiceImpl.java
@@ -1,0 +1,5 @@
+package com.yapp.ios2.fitfty.domain.user;
+
+public class UserServiceImpl implements UserService{
+
+}
