--- src/main/java/com/yapp/ios2/fitfty/interfaces/user/SignInDto.java
@@ -1,23 +1,0 @@
-package com.yapp.ios2.fitfty.interfaces.user;
-
-import javax.validation.constraints.Email;
-import javax.validation.constraints.NotNull;
-import javax.validation.constraints.Size;
-import lombok.*;
-
-@Getter
-@Setter
-@Builder
-@AllArgsConstructor
-@NoArgsConstructor
-@ToString
-public class SignInDto {
-
-    @NotNull
-    @Email
-    private String email;
-
-    @NotNull
-    @Size(min = 3, max = 100)
-    private String password;
-}
